+++ src/components/PostCard.vue
@@ -1,30 +1,56 @@
 <template>
   <div class="card">
-    <router-link :to="{ name: 'post', params: { slug: post.slug } }">
-      <div class="card-image">
-        <figure class="image is-4by3">
+    <div class="card-image">
+      <a
+        :href="post.url"
+        target="_blank">
+        <figure class="image is-3by2">
           <img :src="post.image">
         </figure>
-      </div>
-      <div class="card-content">
-        <div class="content">
-          <p class="title is-4">
-            {{ post.title }}
-          </p>
-          <p>
-            {{ post.publicationDate.toLocaleDateString() }}
-          </p>
-          <p>
-            {{ post.resume }}
-          </p>
+      </a>
+    </div>
+    <div class="card-content">
+      <div class="content">
+        <h3 class="title is-3">
+          {{ post.title }}
+        </h3>
+        <p class="has-text-grey-light">
+          {{ createdAt }}
+        </p>
+        <div class="tags">
+          <div
+            v-for="(tag, index) in post.tags"
+            :key="index"
+            class="tag"
+            :class="'is-' + tag.slug">
+            {{ tag.name }}
+          </div>
         </div>
+        <p>
+          {{ post.subTitle }}
+        </p>
+        <a
+          :href="post.url"
+          class="button is-black is-outlined">
+          <span class="icon">
+            <font-awesome-icon :icon="['fab', 'medium']" />
+          </span>
+          <span>
+            Lire sur Medium
+          </span>
+        </a>
       </div>
-    </router-link>
+    </div>
   </div>
 </template>
 
 <script>
 export default {
-  props: ['post']
+  props: ['post'],
+  computed: {
+    createdAt () {
+      return this.moment(this.post.createdAt).locale('fr').format('dddd Do MMMM YYYY')
+    }
+  }
 }
 </script>