+++ src/store/mutation-types.js
@@ -1,3 +1,4 @@
 export const SET_POSTS = 'SET_POSTS'
 export const SET_RESUME = 'SET_RESUME'
 export const SET_INFO = 'SET_INFO'
+export const SET_DARK_MODE = 'SET_DARK_MODE'