+++ src/views/Contact.vue
@@ -1,6 +1,8 @@
 <template>
   <div>
-    <section class="hero is-dark is-bold">
+    <section
+      class="hero is-bold"
+      :class="modeClass">
       <div class="hero-body has-text-centered">
         <div class="container">
           <h1 class="title is-1">
@@ -9,7 +11,9 @@
         </div>
       </div>
     </section>
-    <section class="section">
+    <section
+      class="section"
+      :class="sectionClass">
       <div class="container">
         <div class="content has-text-centered">
           <p>
@@ -17,6 +21,7 @@
           </p>
           <p>
             <a class="button is-link is-medium"
+              :class="buttonClass"
               :href="'mailto:' + info.mail">
               <span class="icon is-medium">
                 <font-awesome-icon icon="envelope" />
@@ -31,7 +36,9 @@
             ...vous pouvez aussi m'envoyer un messsage sur Twitter :
           </p>
           <p>
-            <a class="button is-link is-medium"
+            <a
+              class="button is-link is-medium"
+              :class="buttonClass"
               href="https://twitter.com/messages/compose?recipient_id=71254404">
               <span class="icon is-medium">
                 <font-awesome-icon :icon="['fab', 'twitter']" />
@@ -56,7 +63,16 @@ import { mapState } from 'vuex'
 export default {
   name: 'Contact',
   computed: {
-    ...mapState(['info'])
+    ...mapState(['info', 'darkMode']),
+    modeClass () {
+      return this.darkMode ? 'is-dark' : 'is-light'
+    },
+    sectionClass () {
+      return this.darkMode ? 'has-background-black has-text-white' : ''
+    },
+    buttonClass () {
+      return this.darkMode ? 'is-light is-inverted' : ''
+    }
   }
 }
 </script>