+++ src/router/index.js
@@ -8,8 +8,7 @@ import Contact from '@/views/Contact'
 Vue.use(Router)
 
 export default new Router({
-  base: process.env.NODE_ENV === 'development' ? '/' : '/romainquencez/',
-  mode: 'history',
+  base: process.env.NODE_ENV === 'development' ? '' : '/romainquencez/',
   routes: [
     {
       path: '/',